Updated backend, added comments schema and started working on friend request.
