+++ backend/routes/api/posts.js
@@ -1,17 +1,14 @@
 const express = require("express");
 const router = express.Router();
 const { check, validationResult } = require("express-validator");
-const Post = require("../../models/Posts");
+const { Comment, Post } = require('../../models/Posts');
 const Profile = require("../../models/Profile");
 const User = require("../../models/User");
 const auth = require("../../middleware/auth");
 
 // Get api/post
 
-router.post(
-  "/",
-  [auth, [check("text", "Text is required").not().isEmpty()]],
-  async (req, res) => {
+router.post("/", [auth, [check("text", "Text is required").not().isEmpty()]], async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json({ errors: errors.array() });
@@ -153,40 +150,5 @@ router.put("/unlike/:id", auth, async (req, res) => {
     res.status(500).send("Server Error");
   }
 });
-
-// POST api/post/comment/:id
-// Comment on a post
-
-router.post(
-  "/comment/:id",
-  [auth, [check("text", "Text is required").not().isEmpty()]],
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    try {
-      const user = await User.findById(req.user.id).select("-password");
-
-      const post = await Post.findById(req.params.id);
-
-      const newComment = {
-        text: req.body.text,
-        name: user.name,
-        user: req.user.id,
-      };
-
-      post.comments.unshift(newComment);
-
-      await post.save();
-
-      res.json(post.comments);
-    } catch (err) {
-      console.error(err.message);
-      res.status(500).send("Server Error");
-    }
-  }
-);
 
 module.exports = router;
